servod: Don't ask user to choose if only 1 servo-v4 + 1 servo-micro
It is a common scenario for a developer connecting one servo-v4 and
one servo-micro (or CCD endpoint). Don't ask user to choose and start
the servo v4. The servo-micro (or CCD) will be initialized in the
post_init phase.

BUG=chromium:665186
TEST=Connected servo v4 and servo-micro (or CCD endpoint). Started
servod and verified no asking.

Change-Id: Ibab126a1b98bd594c19283cdf8d0e3b8bbe0e5e1
Reviewed-on: https://chromium-review.googlesource.com/411499
Commit-Ready: Wai-Hong Tam <waihong@google.com>
Tested-by: Wai-Hong Tam <waihong@google.com>
Reviewed-by: Kevin Cheng <kevcheng@chromium.org>
Reviewed-by: Todd Broch <tbroch@chromium.org>