cros_ec_pd_softrec_power: Fix rec mode (again).
My previous solution to the power_state:rec issue was a bit complex and
caused certain FAFT tests to fail (such as firmware_CorruptBothFwSigAB).
The reason for my previous fix was that there was a race from releasing
cold_reset and issuing the command to the EC to reboot and keep the AP
off.  However, it seems that asserting the cold_reset line isn't
actually necessary and by removing that part, the solution becomes a lot
simpler and the whole "double recovery boot to clear EC SW Sync
failures" is totally unnecessary.

BUG=chromium:648829
BRANCH=None
TEST=Run firmware_CorruptBothFwSigAB against chell and verify that it
passes with a servo v2.
TEST=Run platform_ServoPowerStateController_USBPluggedin against chell
and verify that it passes with a servo v2.
TEST=Run platform_ServoPowerStateController_USBUnplugged against samus
and verify that it passes with a servo v2.
TEST=Repeat the above tests on a servo v3 and verify that they pass.

Change-Id: Ia987c084280c2e69702fe29683b6da8584d9b2d8
Signed-off-by: Aseda Aboagye <aaboagye@google.com>
Reviewed-on: https://chromium-review.googlesource.com/423211
Commit-Ready: Aseda Aboagye <aaboagye@chromium.org>
Tested-by: Aseda Aboagye <aaboagye@chromium.org>
Reviewed-by: Julius Werner <jwerner@chromium.org>
Reviewed-by: Shelley Chen <shchen@chromium.org>