power_state: add ccd_reset command
Servo currently cannot recover any of the stm32 usb endpoints if there
is a usb reset.

For Cr50 testing we may need to reset Cr50 or disconnect CCD
momentarily. This will lead to a USB reset.  This change ccd_reset to
power state which will reinitialize all servo interfaces when called.
For ccd this means that each usb endpoint will be reset.

With this change all interfaces with "reinitialize" methods will
reinitialize when "power_state:ccd_reset" is called. Any reinitialize
methods should be implemented so that they don't break anything if the
interface is already running.

BUG=chrome-os-partner:57468
BRANCH=none
TEST=connect suzyq or servo v4 to dut. start servod using ccd.
disconnect suzyq and reconnect it. run dut-control ccd_reset_usb:on and
then run dut-control and verify all of the fields are populated. verify
all of the consoles work.

Change-Id: I5b71464cef9541edffecb5195df9fb16d00c37ce
Signed-off-by: Mary Ruthven <mruthven@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/445926
Reviewed-by: Wai-Hong Tam <waihong@google.com>