ec: Add pwr_button_hold control
The traditional way to control power button is to set pwr_button to
press and then release. Doing it via an EC console command is a bit
different, as the "powerbtn" command only accepts a hold interval
as its parameter. So add a new control pwr_button_hold for doing it.

BUG=chrome-os-partner:62904
TEST=Started servod and used dut-control to set its value, like:
$ dut-control pwr_button_hold:200
$ dut-control pwr_button_hold:1200
$ dut-control pwr_button_hold:8500

Change-Id: Ib5737324461c104708cd58f8103a1bee1d76ec90
Reviewed-on: https://chromium-review.googlesource.com/443129
Commit-Ready: Wai-Hong Tam <waihong@google.com>
Tested-by: Wai-Hong Tam <waihong@google.com>
Reviewed-by: Wai-Hong Tam <waihong@google.com>