cr50: add control to see if ccd is working
When we are running Cr50 tests, we need a fast way to check if the ccd
endpoints are working. Now 'dut-control ccd_state' will run a cr50
console command with a short timeout to see that the connection is still
up.

BUG=none
BRANCH=none
TEST=start servo and run 'dut-control ccd_state' to see that it returns
on. Reboot cr50 and run 'dut-control ccd_state' and verify it returns
off.

Change-Id: I44ded2a8413157a7abf1eaa239a8f3a2cd26d540
Signed-off-by: Mary Ruthven <mruthven@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/456624
Reviewed-by: Wai-Hong Tam <waihong@google.com>