servo: add reef to list of boards with usbpd uart
Reef boards can get reworked to have the cr50 uart connected to the
servo header. This adds reef to the list of boards with usbpd uart on
interface 0, so you can start servo with 'sudo servod -b reef -c
cr50.xml' and then access the cr50 console.

BUG=none
BRANCH=none
TEST=start servo with 'sudo servod -b reef -c cr50.xml' and run
firmware_Cr50DeepSleepStress on a reef that has the cr50 uart connected
to the servo header.

Change-Id: I31a7913a558c063553fcd380fae69fe3beedf699
Signed-off-by: Mary Ruthven <mruthven@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/469155
Reviewed-by: Aseda Aboagye <aaboagye@chromium.org>