servo: Align the interface index between overlay and internal lists
The interface index defined in the overlay is shifted 1 and different
with the internal list variables, like the control with interface="1",
actually points to the #0 element of the internal lists. It is error
prone if people don't know this shift.

This change aligns the index by adding a dummy #0 element to the lists.

BUG=b:37473185
TEST=Tested on servo v2 and servo v4, verified several interfaces work.

Change-Id: Id63e8c1857e92b187cadb7f9065d556d8a54252d
Reviewed-on: https://chromium-review.googlesource.com/489443
Commit-Ready: Wai-Hong Tam <waihong@google.com>
Tested-by: Wai-Hong Tam <waihong@google.com>
Reviewed-by: Kevin Cheng <kevcheng@chromium.org>