servo: usbpd.xml: Fix servo_v4_interface typo.
There was a typo in the parameter for the raw_usbpd_uart_pty control for
the servo_v4_interface.  This led the control to override the wrong
interface when a servo micro was used behind a servo v4.  The side
effect was that for devices that used the usbpd.xml overlay, the EC
console would be jumbled, which is an issue since it's essential for
power_state functionality.

This commit fixes the typo.

BUG=chromium:714808
BRANCH=None
TEST=Connect a servo micro behind a servo v4.  Run servod with samus.
Verify that EC console works and that output is not garbled.

Change-Id: Ib84cf579690dc3a9340ea2946ad98485cdbd99d7
Signed-off-by: Aseda Aboagye <aaboagye@google.com>
Reviewed-on: https://chromium-review.googlesource.com/489686
Commit-Ready: Aseda Aboagye <aaboagye@chromium.org>
Tested-by: Aseda Aboagye <aaboagye@chromium.org>
Reviewed-by: Kevin Cheng <kevcheng@chromium.org>