servo: cr50: change servo_v4_interface
When setting the control interface the servo_v4_interface will be used
when servo_v4 is doing ccd and when it is using servo micro. This change
redefines the servo_v4_interface in cr50.xml to be the servo micro
interface. It also redefines the whole control in ccd_cr50.xml so that
the servo_v4_interface will be the correct one when using ccd.

BUG=none
BRANCH=none
TEST=start servod -b reef -c cr50.xml using servo v4 with servo micro
and servod -b reef using servo v4 with ccd.

Change-Id: I1661a36c463cf4b796709014a0e21e091e90b662
Signed-off-by: Mary Ruthven <mruthven@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/502189
Reviewed-by: Aseda Aboagye <aaboagye@chromium.org>