servo: Indicate if using CCD or servo micro w/ v4.
It will be helpful to programatically determine what kind of servo
hardware is being used.  This commit appends to the version if a servo
micro is being used or a captive cased closed debug cable in addition to
a servo_v4.

BUG=b:35648297
BRANCH=None
TEST=Run servod with a servo_v4 with both a servo micro and just the CCD
cable.  Verify that the various controls still work.

Change-Id: Ib18c8d8e245bc8723a6e8453e26cc23462194f7f
Signed-off-by: Aseda Aboagye <aaboagye@google.com>
Reviewed-on: https://chromium-review.googlesource.com/502428
Tested-by: Aseda Aboagye <aaboagye@chromium.org>
Reviewed-by: Wai-Hong Tam <waihong@google.com>
Commit-Queue: Aseda Aboagye <aaboagye@chromium.org>