ptydriver: stop using a modular-level variable for _dict
When there are multiple ptyDrivers like ec_uart and cr50_console and a
uart param is set in one driver, then it will also change in the other
driver. This change makes _dict an attribute of _interface to fix this.

This change also modifies ec.py to do the same with the kbd dict.

BUG=none
BRANCH=none
TEST=run firmware_FAFTSetup on a Lumpy and Electro

Change-Id: I19946df1b0d721f1cb7e7c1bd6c56a3ce7ab3418
Signed-off-by: Mary Ruthven <mruthven@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/505408