servo: Add servo micro support for PD MCU UART.
When running servod with a board that has a separate PD MCU UART with a
servo micro, servod would immediately fail when intializing.  This
commit fixes the issue and allows servo_micro to be used with boards
that have a PD MCU.

BUG=None
TEST=Connect servo micro; Verify that servod can be run with both elm
and samus.

Change-Id: Ice0dc3b059eb785358cebfde8fd57c8c9247ddf2
Signed-off-by: Aseda Aboagye <aaboagye@google.com>
Reviewed-on: https://chromium-review.googlesource.com/583712
Commit-Ready: Aseda Aboagye <aaboagye@chromium.org>
Tested-by: Aseda Aboagye <aaboagye@chromium.org>
Reviewed-by: Mary Ruthven <mruthven@chromium.org>