servo: Add get_serial_name() method.
This commit adds a method that will allow the various connected servos' serial
numbers to be retrieved.

BUG=chromium:747112
TEST=sudo emerge hdctools; `dut-control serialname` returns the main
servo serial number.  With a servo v4 with a servo micro connected to
it, verify that `dut-control servo_micro_serialname` retrives the serial
number of the servo micro.

Change-Id: I15f03dfe56f0becc549fa7258b102982605465fb
Signed-off-by: Aseda Aboagye <aaboagye@google.com>
Reviewed-on: https://chromium-review.googlesource.com/597209
Commit-Ready: Aseda Aboagye <aaboagye@chromium.org>
Tested-by: Aseda Aboagye <aaboagye@chromium.org>
Reviewed-by: Mary Ruthven <mruthven@chromium.org>