servo: cros_ec_softrec: Enable DR swap for servo v4.
For devices using servo_v4 with the captive CCD cable, we need to ensure
that the DUTs port is a downstream facing port "DFP" in order for the
USB ethernet, USB port, and microSD port are seen by the DUT.  Normally,
this happens automatically, but in recovery it does not because EC_RO
normally prevents data role (DR) swap.  This is problematic when the DUT
is trying to boot a recovery image from USB.

This commit triggers a data role swap when `power_state:rec` is called.

BUG=None
TEST=`dut-control power_state:rec` on a DUT with a servo_v4 with captive
Type-C cable and USB flash drive plugged in.  Verify that DUT boots from
USB.

Change-Id: Iaa5c59eda8d4ddbef04b67e034015a118112b392
Signed-off-by: Aseda Aboagye <aaboagye@google.com>
Reviewed-on: https://chromium-review.googlesource.com/620273
Commit-Ready: Aseda Aboagye <aaboagye@chromium.org>
Tested-by: Aseda Aboagye <aaboagye@chromium.org>
Reviewed-by: Nick Sanders <nsanders@chromium.org>