servo: Add servo_v4_ccd_keepalive control.
When servo_v4 is used for CCD, it monitors the SBU lines to determine
which orientation to set its mux.  However, servo_v4 needs to enter a
"keepalive" state to avoid accidentally terminating the CCD connection
during events such as an EC reboot which may appear as disconnects.

This commit adds the control and initalizes the keepalive state to
enabled.

BUG=b:64903997
TEST=Use updated servo_v4 firmware which supports keepalive.  Verify
that `dut-control servo_v4_ccd_keepalive` can retrieve the keepalive
state and can set it as well.
TEST=Additionally, verify that the keepalive state is initialzed to true.

Change-Id: I2ff9eb429084109ff5c09491195e0ad30b7ce114
Signed-off-by: Aseda Aboagye <aaboagye@google.com>
Reviewed-on: https://chromium-review.googlesource.com/648251
Commit-Ready: Aseda Aboagye <aaboagye@chromium.org>
Tested-by: Aseda Aboagye <aaboagye@chromium.org>
Reviewed-by: Mary Ruthven <mruthven@chromium.org>