servo: Introduce get_base_board() to get the board name of the base
Similar to get_board() which returns the board name of the main device,
the get_base_board() returns the board name of the base device if it is
present. It is useful to check if a base is connected and talk to its
related controls.

BUG=b:66211919
TEST=Started servod with the setup connecting to the base. Called the
RPC get_base_board() to its board name.

Change-Id: If1335c4b7834d1c884e19693af13a71e7f2b5162
Reviewed-on: https://chromium-review.googlesource.com/675746
Commit-Ready: Wai-Hong Tam <waihong@google.com>
Tested-by: Wai-Hong Tam <waihong@google.com>
Reviewed-by: Aseda Aboagye <aaboagye@chromium.org>