keyboard: Send commands to BaseEC to emulate key-press if the base is connected
This change is for detachable devices. Traditionally, the key-press emulation
can be done by sending console commands to the main EC. But for detachable
devices, the main EC is not able to do so. The keyboard function is controlled
by the base EC. Should change to redirect the console commands to the BaseEC.

BUG=b:66211919
TEST=Started servod with the setup of servo v4 connecting two micro-servos
to Soraka and Staff. Called the RPCs enter_key() and ctrl_d() to verify.

Change-Id: Id0723014365883d0abe498b6f752ff85620d4b84
Reviewed-on: https://chromium-review.googlesource.com/675747
Commit-Ready: Wai-Hong Tam <waihong@google.com>
Tested-by: Wai-Hong Tam <waihong@google.com>
Reviewed-by: Aseda Aboagye <aaboagye@chromium.org>
Reviewed-by: Nicolas Boichat <drinkcat@chromium.org>