grunt: servo overlay
This gives us some initial settings for servo for Grunt.

This reconfigures the warm_reset pin, as it's now on pin 39 (as it is on
some other new boards).

It also sets the Vref for the PCH SPI write-protect to 3.3V since that's
the voltage everything else seems to use for signaling there.

BUG=b:67187156
TEST=sudo emerge hdctools

Change-Id: I54ce38975c6559876cbf35acbf9ada5b9dbf2953
Reviewed-on: https://chromium-review.googlesource.com/818302
Commit-Ready: Justin TerAvest <teravest@chromium.org>
Tested-by: Justin TerAvest <teravest@chromium.org>
Reviewed-by: Aaron Durbin <adurbin@chromium.org>
Reviewed-by: Justin TerAvest <teravest@chromium.org>