cros_ec_softrec_power: Don't warm reset if it can't hold AP
The procedure of power_state:rec does a warm_reset:on to hold AP before
doing the whole system reboot (cold_reset/PMIC-reset) and setting EC
hostevent for recovery.

However, it doesn't work on some new x86 devices. Even warm_reset is
ON; AP still powers on and executes the verified-boot sequence, i.e.
 * (vboot) Read the recovery reason from NV storage (e.g. 19 for the
   test firmware_CorruptBothFwSigAB, as it failed to verify the fw
   signature in the previous boot);
 * (vboot) Clear the recovery reason in NV storage, such that it will
   do clear normal boot next time;
 * (servod) Servod performs the whole system reboot and sets the EC
   hostevent for recovery.
 * (vboot) Read the recovery reason from NV storage, i.e. 0;
 * (vboot) As no recovery reason, treats it as manual-triggered
   recovery (recovery reason: 2).

So some of the FAFT tests failed, as the expected recovery reason is
not matched, 19 -> 2.

A fix to this race condition is not to do warm reset for the devices
can't hold AP. For devices having the recent EC patch of the command
'reboot wait-ext' command, it can hold AP off, without the warm_reset.

BUG=b:70573209
TEST=Ran firmware_CorruptBothFwSigAB and firmware_RollbackFirmware
passed on Fizz.

Change-Id: Ia8f8ac11704e5c31e195ec6a85665102634a648c
Reviewed-on: https://chromium-review.googlesource.com/830253
Commit-Ready: Wai-Hong Tam <waihong@google.com>
Tested-by: Wai-Hong Tam <waihong@google.com>
Reviewed-by: Randall Spangler <rspangler@chromium.org>
Reviewed-by: Scott Collyer <scollyer@chromium.org>