servo: refactor generate_ina_controls.py
Making generate_ina_controls.py more modular, and removing
large hard-coded strings. This is to allow for easier extensions,
and improve readability on what ina controls are being generated.

This also introduces a new method to generate multiple, identical INA
files used on different revisions. For instance, poppy rev 1 and rev 2 have
the same INA map. To make boostrapping servod easier, this
change now allows for a |revs| attribute inside the
[inas].py files.
|revs| is a list of revisions where this INA map
applies to. See poppy_2.py for an example.

Also move logic out of Makefile, and into setup.py. This way simply
adding a new .py file with an inas attribute will try to generate
new controls. No need to explicitly add it in the Makefile.

Lastly, we've had issues before with the INA .py file formatting
so this includes a helper to generate a Skeleton configuration file
and help to make new files.

BUG=chromium:760267
TEST=manual testing
- generate poppy_r3.xml with new file
- generate poppy_r3.xml with old file
diff & see that the only difference is the order of some attribtues.

Change-Id: I479aabc0e313adc39a30de3e07ae1d8060028069
Reviewed-on: https://chromium-review.googlesource.com/848873
Commit-Ready: Ruben Rodriguez Buchillon <coconutruben@chromium.org>
Tested-by: Ruben Rodriguez Buchillon <coconutruben@chromium.org>
Reviewed-by: Ruben Rodriguez Buchillon <coconutruben@chromium.org>