servo: leverage pwr_avg where available
pwr_avg was recently introduced as an ec command that provides a
one minute averaged reading of current, voltage, and power. This
change introduces the ec code and xml files required to leverage
this. Now there is:
- avg_ppvar_vbat_mw
- avg_ppvar_vbat_ma
- avg_ppvar_vbat_mv

Since the command is fairly new, the code also explicitly checks
for the command being available, and returns a more informative error
if the failure is due to pwr_avg not being available on that ec image.
Any code, such as power scripts, can then decide higher up in the stack
if pwr_avg being unavailable is a deal breaker or if it's okay to
default to ppvar_vbat_mw.

BUG=chromium:752320, chromium:760267
TEST=manual
$ dut-control avg_ppvar_vbat_mw
avg_ppvar_vbat_mw:5572
(on dut without the command)
$ dut-control avg_ppvar_vbat_mw
Problem with ['avg_ppvar_vbat_mw'] :: cmd |pwr_avg| is not available on
the ec.

Change-Id: I4b9efeaa87a5c5f6ca932cfa2cec7cb135f7d388
Reviewed-on: https://chromium-review.googlesource.com/888281
Commit-Ready: Ruben Rodriguez Buchillon <coconutruben@chromium.org>
Tested-by: Ruben Rodriguez Buchillon <coconutruben@chromium.org>
Reviewed-by: Todd Broch <tbroch@chromium.org>