ftdiuart: Put in an explicit delay for sending a break
Recently after switching to a new desktop, Linux version, and
reogranizing my USB devices (not sure which of the three was
important), I found that I could no longer manage to send a break
through servo to cause a SYSRQ on Linux.

Although I didn't scope the signal, I postulated that perhaps the
break signal was not being held long enough.  After all, the API to
libftdi just allows you to start asserting a break and then stop.  It
doesn't appear to have any sort of timing built in.  Sure enough,
adding a usleep() between asserting the break and deasserting it fixed
my problem.

From some quick searching, it appears that you're supposed to hold the
break state for "longer than one frame time".  That's pretty easy to
calculate since a frame is made up of:
- 1 start bit
- 7 or 8 data bits
- 0 or 1 parity bits
- 1, 1.5, or 2 stop bits
...or, at most 12 bits.

Apparently there are separate concepts for "short break" and "long
break" but we don't really have the ability to get such precise
timing so we'll just make sure that we at least qualify for some type
of break and call it a day.

BUG=None
TEST=Type "<enter>`zg" and enter kgdb

Change-Id: I40376afe6f466ff8112cfc6276a29fcf2f8f8960
Reviewed-on: https://chromium-review.googlesource.com/962672
Commit-Ready: Douglas Anderson <dianders@chromium.org>
Tested-by: Douglas Anderson <dianders@chromium.org>
Reviewed-by: Todd Broch <tbroch@chromium.org>