servod: set safe default for shutdown delay
power_state:off has a delay to let the system settle,
but the default is extremely aggrerssive and causes failures on
Dru.

Change the default to match the conservative side of board overrides,
and change Dru to specify a specific value since it's empirically needed.

BUG=b:76229658
TEST=dut-control power_state:off && dut-control power_state:on

Change-Id: I4a96b02b0327ef856f70141ee46e1add9eacba2a
Reviewed-on: https://chromium-review.googlesource.com/979260
Commit-Ready: Nick Sanders <nsanders@chromium.org>
Tested-by: Nick Sanders <nsanders@chromium.org>
Reviewed-by: Aseda Aboagye <aaboagye@chromium.org>