servo: add configuration integration test
This generates a host of tests to validate the integrity of servo/data
configuration files.
The tests validate each control, map, and include being structured
properly, as well as some cross-reference validation, such as making
sure that maps used are defined somewhere, and include files exist.

Note: the current 4 failures are due to power XML files being imported,
but those only being generated in the build stage.
A TODO is left in data_integrity_test to allow to run outside of the
ebuild by builing the data-package in a temporary directory.
However, this does not break the FEATURES=test flow.

BUG=chromium:812152
TEST=manual
py.test data_integrity_test.py
...
4 failed, 8153 passed, 3499 skipped in 13.77 seconds

Change-Id: I2db9dcd52af8623071588d91dab663062671a07f
Signed-off-by: Ruben Rodriguez Buchillon <coconutruben@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/1078588
Reviewed-by: Todd Broch <tbroch@chromium.org>