servod: fail on lack of device interface
Currently servod will start with only servo v4 available,
and merely fail later with unattached device controls, if a servo
micro or CCD interface isn't found.

If a board is specified, require that a board interface (that is,
servo micro or CCD) is found to enable the standard controls requested
in the board overlay xml.

BUG=b:79550986
TEST=error when no servo or ccd is connected.

Change-Id: Ifa75071488b7bb5456bcbf1f4f9761918670eec9
Signed-off-by: Nick Sanders <nsanders@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/1097809
Reviewed-by: Wai-Hong Tam <waihong@google.com>
Reviewed-by: Mary Ruthven <mruthven@chromium.org>