octopus: Add INAs config for newer boards.
Yorp 1.4 (and boards based off that design) have a different resistor
value for the sense resistor for PP3300_EC. This commit creates a new
revision for the INA configuration, and removes the default from the
octopus_npcx overlay.

We'll update servod usage instructions for Octopus accordingly with
instructions for choosing the right xml file for INA configuration.

BUG=b:80490867
TEST=None

Change-Id: Ibbb5e9ee648aeda1f2bd53bb3b61f6a540698024
Reviewed-on: https://chromium-review.googlesource.com/1099609
Commit-Ready: Justin TerAvest <teravest@chromium.org>
Tested-by: Justin TerAvest <teravest@chromium.org>
Reviewed-by: Furquan Shaikh <furquan@chromium.org>
Reviewed-by: Jett Rink <jettrink@chromium.org>