servod: add usb image management logic
This is the first CL in a sequence of CLs (that are to be considered one
change really) to move the usb image management logic into a driver.

This CL introduces two common aliases on V2 and V4:
image_usbkey_mux
image_usbkey_pwr

It also introduces the usb_image_management driver that can safely
switch the direction of the image_usbkey by power-cycling before
switching.

BUG=chromium:869335
TEST=works on v2, v3 & v4 as expected

Change-Id: Ia242143f5b859bf8b68e264a7edd519b25d4e878
Reviewed-on: https://chromium-review.googlesource.com/1162097
Commit-Ready: Todd Broch <tbroch@chromium.org>
Tested-by: Ruben Rodriguez Buchillon <coconutruben@chromium.org>
Reviewed-by: Todd Broch <tbroch@chromium.org>