servod: introduce servodutil to manage servod instances easier.
This change introduces servodutil.py. The main function it provides is a
tool to leave behind breadcrumbs when starting a servod instance that
make it possible to
(1) gently turn down a servod instance
(2) get an overview of the servod instances running on a workstation

Usage examples (assume a servod instance being served on port 9999)
$ servodutil show -p 9999
port: 9999
serials = ['someserial']
pid: xxxx
$ sudo servodutil stop -p 9999
SIGTERM sent to servod instance associated with '9999'.
$ servodutil show-all
No entries found.

BUG=chromium:841121
TEST=manual testing, and py.test -v servodutil_test.py

Change-Id: Ia9b74d3b3b19254b2145b523315e6c9101791a05
Signed-off-by: Ruben Rodriguez Buchillon <coconutruben@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/1117748