servod: small measure_power refactor
This refactor gets rid of the split between PowerTracker and
ServodPowerTracker.
Originally the motivation here had been to allow for flexibility in also
having a powerlog power-tracker. But since our current efforts are on
unifying behind the servod interface, this is an unnecessary
modularization.
Additionally this CL improves some documentation that was lacking.

BUG=chromium:806146
TEST=manual testing
dut-power -t 120

Change-Id: I8f44aa544ac030de890658120f729f705269bfc5
Signed-off-by: Ruben Rodriguez Buchillon <coconutruben@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/1229534
Reviewed-by: Mengqi Guo <mqg@chromium.org>