servod: use avg_ppvar in measure_power.py
This change leverages avg_ppvar in the measure_power script.
It does so by ensuring that the ECPowerTracker always runs the first
sample with normal ppvar_vbat_mw (to not get the baggage of the previous
minute) and then switches to avg_ppvar_vbat_mw if that is available on
the DUT's EC image.

BUG=chromium:806146, chromium:867383
TEST=manual testing
dut-control -t 130 on a dut with avg_ppvar enabled.

Change-Id: I6abf4aebd0eea236fc42878c89c852990e939281
Signed-off-by: Ruben Rodriguez Buchillon <coconutruben@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/1229535
Reviewed-by: Mengqi Guo <mqg@chromium.org>