servod: use dsleep in measure_power to increase ec vbat accuracy
This adds a prepare stage to the ECPowerTracker to issue a dsleep
control to the EC console. This allows the EC (if dsleep is supported)
to enter dee-sleep after 2s and not the standard 15. This greatly
improves accuracy, especially in suspend/power-off measurement
scenarios.

BUG=chromium:806146
TEST=manual testing
control gets sent to the EC successfully

Change-Id: Iaffa88e17acfae70d7e826bc7ac98dea6aeb47b2
Reviewed-on: https://chromium-review.googlesource.com/1229536
Commit-Ready: ChromeOS CL Exonerator Bot <chromiumos-cl-exonerator@appspot.gserviceaccount.com>
Tested-by: Ruben Rodriguez Buchillon <coconutruben@chromium.org>
Reviewed-by: Puthikorn Voravootivat <puthik@chromium.org>