servod: building a parser with native rc support
This change pulls the servo rc logic into a ServodRCParser, so that
clients of the RC logic need no knowledge of it. It also changes
the RC logic to discontinue using the |port| in the servodrc files, but
instead leverage ServoScratch to find what port a serialname is running
under. For now, the rc-file format will not change, but eventually, as
port is discontinued, an update + PSA, or some "smart" logic to detect
if the 2nd entry in the rc is a port, or a board-file will be needed.

BUG=chromium:841121
TEST=manual testing

servod:
-n works natively
-n & -s raises an error
-s works fine

dut-control/dut-power:
-n/-s leverage the scratch to find the port
-p & -n/-s successfully prefers -p over the others

Cq-Depend: chromium:1369464
Change-Id: Id0b5ae38cb61f300c013a10e37f0f63ede69f099
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/hdctools/+/1232800
Tested-by: Ruben Rodriguez Buchillon <coconutruben@chromium.org>
Commit-Queue: Ruben Rodriguez Buchillon <coconutruben@chromium.org>
Reviewed-by: Puthikorn Voravootivat <puthik@chromium.org>
Reviewed-by: Todd Broch <tbroch@chromium.org>