servod: add -m model arg
This adds an optional argument '-m' which specifies
a model under a board. Starting with Octopus, not all
devices within a reference design/overlay share the
same servo header and component layout. We do currently
guarantee that model specifies a particular MLB, so
this is a sufficient and discoverable disambiguation.

BUG=chromium:886955
TEST=the right files load for phaser and apel.

Change-Id: I5946b3ce3f87c4762d6650c7affb9121fe675d27
Reviewed-on: https://chromium-review.googlesource.com/1235334
Commit-Ready: Nick Sanders <nsanders@chromium.org>
Tested-by: Nick Sanders <nsanders@chromium.org>
Reviewed-by: Justin TerAvest <teravest@chromium.org>