servod: system_config support printing prefix in display_config
For multi device on servod support, we need to be able to print accurate
configurations. However, the servo device class is to be shielded from
knowing anything about prefixes, or routing.
This CL allows system_config to print all its controls with a prefix
string - the prefix used to route controls to its servo device.

BUG=chromium:841121
TEST=(with subsequent CLs)
inspect that all controls with a prefix are printed with
[prefix].control

Change-Id: Icf86fb2ca1d8cdf90aec3f286312bdb00a3cc944
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/hdctools/+/1242501
Tested-by: Ruben Rodriguez Buchillon <coconutruben@chromium.org>
Commit-Queue: Ruben Rodriguez Buchillon <coconutruben@chromium.org>
Reviewed-by: Todd Broch <tbroch@chromium.org>
Reviewed-by: Puthikorn Voravootivat <puthik@chromium.org>
Reviewed-by: Brian Nemec <bnemec@chromium.org>