servod: increase watchdog polling rate for ccd to 100ms
There are reports that the watchdog might miss a cr50-reboot if it's
triggered through the cr50 console instead of powerstate driver.
This aims to mitigate that by increasing the polling rate for the
watchdog from 1s to 100ms if a ccd device is part of servod.

BUG=b:117323430
TEST=manual testing
sudo servod -b soraka &
... - ... - INFO - Reinit capable device found. Polling rate set to 0.10s.

Change-Id: I0cdaa6778c98e084e09b5d000b1ff25cd1cfaff9
Reviewed-on: https://chromium-review.googlesource.com/1267855
Commit-Ready: ChromeOS CL Exonerator Bot <chromiumos-cl-exonerator@appspot.gserviceaccount.com>
Tested-by: Ruben Rodriguez Buchillon <coconutruben@chromium.org>
Reviewed-by: Mary Ruthven <mruthven@chromium.org>