servod: don't mask errors in ec3po interpretation.
This code is masking each timeout by returning res, which is None in
case the issue_cmd fails. This should fix this, so that at least we see
which errors are actually due to timeouts (i.e. console inavailability).

BUG=b:117848487
TEST=Ran servod in an environment with a timeout error.  When combined
with CL:1286216 to not suppress tracebacks further up the stack, saw the
traceback on stderr.

Change-Id: Ic0f0514c4049bb23c468670a4b69cbb931403826
Signed-off-by: Ruben Rodriguez Buchillon <coconutruben@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/1286589
Commit-Ready: ChromeOS CL Exonerator Bot <chromiumos-cl-exonerator@appspot.gserviceaccount.com>
Tested-by: Matthew Blecker <matthewb@chromium.org>
Reviewed-by: Matthew Blecker <matthewb@chromium.org>