servod: not_applicable fixup
In a recent change to the NA driver, the 'na' map & subtype were
removed. However, autotest code relies on an na driver returning
'not_applicable' instead of the 0 it returns now. There's no difference
from a servod perspective what to return, so to keep changes to a
minimum, this adds back the functionality to return 'not_applicable' on
a get call to an na driver. This is not done with a map, but rather just
as a default

BUG=b:118854546
TEST=manual testing
sudo servod -b fizz
dut-control lid_open
> lid_open:not_applicable

Change-Id: I4843ab56d84410071712ee6c963060470cd5f53a
Signed-off-by: Ruben Rodriguez Buchillon <coconutruben@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/1314270
Reviewed-by: Shelley Chen <shchen@chromium.org>