usbkm: Set fixed endpoint size to 64
The current device descriptor sets the default control endpoint size
to a fixed 8 bytes.  This does not work in the depthcharge xhci driver
which is not currently able to split control transactions for basic
operations like 'get descriptor'.

In order to make this keyboard work in depthcharge set the fixed
control endpoint size to 64 bytes. (valid values are 8, 64, 512)

BUG=b:121380403
TEST=pass firmware_DevBootUSB on sarien

Change-Id: Id878859484de66ca4cddb85de234d8878494189c
Signed-off-by: Duncan Laurie <dlaurie@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/1415843
Commit-Ready: Duncan Laurie <dlaurie@google.com>
Tested-by: Duncan Laurie <dlaurie@google.com>
Reviewed-by: Nick Sanders <nsanders@chromium.org>