sarien: Update sarien power driver
The recovery mode boot was doing two cold resets which is not
necessary and may be causing issues with some power behavior.
Also ensure that the recovery mode boot waits for the configured
boot screen delay before returning, but do not wait 5 seconds
after asserting the recovery mode pin to power on as that is
an unnecessary delay.

BUG=b:120958278
TEST=platform_ServoPowerStateController.USBPluggedin

Change-Id: I4d581636346b811f31b6c15b666f9fd8f66d806d
Signed-off-by: Duncan Laurie <dlaurie@google.com>
Reviewed-on: https://chromium-review.googlesource.com/1529044
Reviewed-by: Furquan Shaikh <furquan@chromium.org>