zerblebarn: disconnect cpu ec3po
The fpmcu uart on zerblebarn is connected to uart2. Disconnect the cpu
uart from uart2, so the fpmcu ec3po can connect to uart2. This change
turns cpu_ec3po_interp_connect off and changes the map, so it can't be
turned on.

BUG=none
BRANCH=none
TEST=start servod with the zerblebarn overlay. Make sure the fpmcu uart
gets all of the uart output and the cpu uart doesn't do anything.

Change-Id: I0b90a6de847073d3a761c8f2b5c8401a610e4446
Signed-off-by: Mary Ruthven <mruthven@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/1597429
Reviewed-by: Namyoon Woo <namyoon@chromium.org>