Servod: Added handling of additional IPC signals for proper shutdown
Servod requires a shutdown sequence to be executed for it to cleanly stop
in a known good configuration. If the process terminates in a way which
is not handled then servod may not be able to resume until the appropriate
recovery path is taken. This recovery path may require involve the host
system having resources locked resulting in 'Resource Busy' errors.
External programs which communicate with the servod process can stall
when timeouts are missing. If the Servo or the DUT is busy responding
to commands, it may not be in a state which permits it to respond
properly to the unexpected initialization sequence so physically
disconnecting and power cycling components may be needed as recovery. This
does not fully resolve that situation but seems to reduce the rate of
occurrence.

BRANCH=servo
BUG=b:134093234
TEST=Verified servod and servo recovers reliably from a variety of
shutdown conditions including closing terminal and sending various
classes of keyboard signals including SIGINT, SIGTSTP, and SIGQUIT
Signed-off-by: Brian Nemec <bnemec@chromium.org>

Change-Id: I6950365caa799a95698fb37083f92ef5da095cf2
Reviewed-on: https://chromium-review.googlesource.com/1636862
Commit-Ready: Brian Nemec <bnemec@google.com>
Tested-by: Brian Nemec <bnemec@chromium.org>
Legacy-Commit-Queue: Commit Bot <commit-bot@chromium.org>
Reviewed-by: Todd Broch <tbroch@chromium.org>